Rename package (#1212)
