+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/envelopehandlers/SupplementaryEvidenceWithOcrHandler.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.events;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.envelopehandlers;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.PaymentsProcessor;
@@ -10,12 +10,12 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.processedenvelopes.EnvelopeCcdAction.EXCEPTION_RECORD;
 
 @Service
-public class ExceptionClassificationHandler {
+public class SupplementaryEvidenceWithOcrHandler {
 
     private final CreateExceptionRecord exceptionRecordCreator;
     private final PaymentsProcessor paymentsProcessor;
 
-    public ExceptionClassificationHandler(
+    public SupplementaryEvidenceWithOcrHandler(
         CreateExceptionRecord exceptionRecordCreator,
         PaymentsProcessor paymentsProcessor
     ) {
@@ -25,8 +25,8 @@ public class ExceptionClassificationHandler {
 
     public EnvelopeProcessingResult handle(Envelope envelope) {
         checkArgument(
-            envelope.classification == Classification.EXCEPTION,
-            "Envelope classification has to be " + Classification.EXCEPTION
+            envelope.classification == Classification.SUPPLEMENTARY_EVIDENCE_WITH_OCR,
+            "Envelope classification has to be " + Classification.SUPPLEMENTARY_EVIDENCE_WITH_OCR
         );
 
         Long ccdId = exceptionRecordCreator.tryCreateFrom(envelope);
